Revert "[export][refactor][2/n] Move tracing logic (#114768)"
This reverts commit 0ab57ee7eab5391289d30e8c49fceee3f503f539.
