+++ test/export/test_pass_infra.py
@@ -5,7 +5,7 @@ from typing import List
 import torch
 from functorch.experimental import control_flow
 from torch._dynamo.eval_frame import is_dynamo_supported
-from torch.export import export
+from torch._export import export
 from torch._export.pass_base import _ExportPassBase
 from torch.testing._internal.common_utils import run_tests, TestCase
