+++ torch/export/__init__.py
@@ -170,7 +170,7 @@ def export(
       ``OrderedDict`` containing all above types.
 
     """
-    from ._trace import _export
+    from torch._export import _export
     from .dynamic_shapes import _process_dynamic_shapes
 
     if constraints is not None:
